requirements.txt
asgiref==3.8.1
blinker==1.8.2
click==8.1.7
colorama==0.4.6
Django==5.1.3
Flask==3.0.3
Flask-MySQLdb==2.0.0
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
pillow==11.0.0
platformdirs==4.2.2
pycodestyle==2.12.1
PyMySQL==1.1.1
sqlparse==0.5.1
tomli==2.0.1
typing_extensions==4.12.2
tzdata==2024.2
Werkzeug==3.0.4
yapf==0.40.2
zipp==3.19.2
